Image_Histograms_SV- Description
This program takes in an image as input and applies a function to it to create an image histogram. It creates a mask to place on the image before deriving an image histogram for the pixels in the mask and displaying both the masked region and the derived image histogram.