Matching_Histograms_SV- Description
This program takes in two images as input and matches them together using their histograms to derive a new image. This can be used to change the background of an image given a separate image where the correct background is included, even if an unpreferred object is present in the other image.