Color_Correction_SV- Description
This program takes in a source and reference ArUCo images and corrects the color on the source image using histogram matching and color correction. The program will create a new image with only the source image, deleting the background, and will detect the ArUCo markers on the image before displaying the images to the screen.